split inbox to its own file
